Serve community builds from default GitHub branch
When Nextstrain community builds were implemented, the default GitHub branch was "master", and this assumption was build into our server. The ability to explicitly define which GitHub branch should be used was possible (by specifying it in the URL via "@branch`), but in the absence of this "master" was always used.

It is now common for the default / only branch to be "main", and it is surprising for users when Nextstrain community builds result in an error in these cases.

Here we query GitHub for the default branch and use that. If the branch is made explicit via the use of `@branch` in the URLs, then that branch is used (unchanged behaviour). Such a query necessitates that rather a lot of methods & functions become async, and some getters have become (normal) functions.