gatsby page for /groups/:groupName ;
this uses some recent work on the static site
to render splash pages for nextstrain groups
using the gatsby server and the new dataset
select interface.

Still to be addressed:
- Parsing & rendering custom markdown from each
group instead of displaying the markdown raw.

Related but maybe in a subsequent PR:
- Expand the express middleware (that guesses whether to
route urls to auspice or gatsby) to check existence of
groups datasets.

- Use the /groups/:groupName page to handle missing
groups datasets urls and show the appropriate group
page with a message about missing dataset.